2019-07-30/live_code/kickstarter/requirements.txt
django
djangorestframework
django-cors-headers

